Add files via upload
This figure is the final SHAP analysis output of the test code. We have the Feature importance ranking in the Y axis, and SHAP values in X axis. The smaller the SHAP value, more reduction in total emissions; whilst red dots meaning high values of that feature and blue dots meaning low values of that feature.

In this figure, Unit Cost Difference is the biggest driver for carbon emissions. A potential insight for this, could be that the cost competitiveness of green resources is a significant factor in determining carbon emissions.